Update steering works_limits
thismostly works, but moreso just for demos. when i decide what to do for the remote control, ill probably have to update it. current issue s that the limits drift bc this is a continuous servo not the positional one. but if i have a potentiometer then that's not as much of an issue. ill just need to create more solid limits, perhaps with solid physical barriers or some other sensors